First stage split of processor into different modules
