+++ src/mireport/xlsx_template_reader/_util.py
@@ -1,9 +1,14 @@
+from __future__ import annotations
+
+import logging
 import re
 import warnings
 from datetime import date, datetime, time
 from pathlib import Path
 from typing import (
+    TYPE_CHECKING,
     BinaryIO,
+    Iterable,
     Iterator,
     Literal,
     NamedTuple,
@@ -13,6 +18,10 @@ from typing import (
     overload,
 )
 
+if TYPE_CHECKING:
+    from mireport.taxonomy import Concept
+
+from dateutil.parser import parse as parse_datetime
 from openpyxl import Workbook, load_workbook
 from openpyxl.cell import Cell, MergedCell, ReadOnlyCell
 from openpyxl.utils.cell import absolute_coordinate, quote_sheetname
@@ -20,13 +29,25 @@ from openpyxl.workbook.defined_name import DefinedName
 from openpyxl.worksheet.cell_range import CellRange
 from openpyxl.worksheet.worksheet import Worksheet
 
+from mireport.conversionresults import ConversionResultsBuilder, MessageType, Severity
 from mireport.exceptions import OpenPyXlRelatedException
 from mireport.typealiases import DecimalPlaces
+from mireport.xlsx_template_reader._bindings import (
+    CellAndXBRLMetadataHolder,
+    CellRangeMetadata,
+)
+
+L = logging.getLogger(__name__)
 
 CellType: TypeAlias = ReadOnlyCell | MergedCell | Cell
 CellValueType: TypeAlias = bool | float | int | str | datetime | date | time | None
 
 EXCEL_PLACEHOLDER_VALUE = "#VALUE!"
+EXCEL_VALUES_TO_BE_TREATED_AS_NONE_VALUE = frozenset({"-", EXCEL_PLACEHOLDER_VALUE})
+
+
+def conceptsToText(concepts: Iterable[Concept]) -> str:
+    return ", ".join(sorted(str(c.qname) for c in concepts))
 
 
 class NamedRangeException(OpenPyXlRelatedException):
@@ -291,6 +312,21 @@ class CellRangeDimensions(NamedTuple):
         return len(self.cellsPopulated)
 
 
+def getDateFromValue(value: object) -> date:
+    if isinstance(value, datetime):
+        return value.date()
+    elif isinstance(value, date):
+        return value
+    elif isinstance(value, str):
+        if "-" in value:
+            return date.fromisoformat(value)
+        elif "/" in value:
+            return parse_datetime(value, yearfirst=False, dayfirst=True).date()
+        raise ValueError(f"Unsupported date string: '{value}'")
+    else:
+        raise TypeError(f"Unsupported type for date conversion: {type(value).__name__}")
+
+
 def getEffectiveCellRangeDimensions(
     ws: Worksheet, cell_range: CellRange
 ) -> CellRangeDimensions:
@@ -334,3 +370,213 @@ def getEffectiveCellRangeDimensions(
         cellsAccessed=cellCount,
         cellsPopulated=populatedCellCount,
     )
+
+
+class WorkbookReader:
+    """Ergonomic cell-level access to an openpyxl Workbook.
+
+    Carries the three shared context items (workbook, unused-name set,
+    results builder) so call sites don't repeat them on every helper call.
+    """
+
+    def __init__(
+        self,
+        workbook: Workbook,
+        unused: set[DefinedName],
+        results: ConversionResultsBuilder,
+    ) -> None:
+        self._workbook = workbook
+        self._unused = unused
+        self._results = results
+
+    def close(self) -> None:
+        self._workbook.close()
+
+    def getDefinedName(self, name: str) -> Optional[DefinedName]:
+        return self._workbook.defined_names.get(name)
+
+    def _getCellRange(self, dn: DefinedName) -> Optional[CellRangeMetadata]:
+        all_destinations = list(dn.destinations)
+        match len(all_destinations):
+            case 0:
+                self._results.addMessage(
+                    f"Named range {dn.name} has no destinations specified. Ignoring.",
+                    Severity.ERROR,
+                    MessageType.DevInfo,
+                )
+                return None
+            case 1:
+                pass
+            case _:
+                self._results.addMessage(
+                    f"Table {dn.name} has multiple destinations. Ignoring table.",
+                    Severity.ERROR,
+                    MessageType.DevInfo,
+                )
+                return None
+        sheetName, cell_range = all_destinations[0]
+        if not sheetName or not cell_range:
+            self._results.addMessage(
+                f"Named range {dn.name} has damaged cell reference {sheetName=} {cell_range=}",
+                Severity.ERROR,
+                MessageType.ExcelParsing,
+            )
+            return None
+        try:
+            ws = self._workbook[sheetName]
+            cr = CellRange(cell_range)
+        except Exception as e:
+            L.exception("OpenPyXL is sad.", exc_info=e)
+            return None
+        dims = getEffectiveCellRangeDimensions(ws, cr)
+        self._results.addCellQueries(dims.cellsAccessed)
+        self._results.addCellsWithData(dims.cellsPopulated)
+        return CellRangeMetadata(
+            dn,
+            ws,
+            cr,
+            effectiveHeight=dims.height,
+            effectiveWidth=dims.width,
+            cellsPopulated=len(dims.cellsPopulated),
+        )
+
+    def getSingleCell(
+        self,
+        definedName: DefinedName | str | CellAndXBRLMetadataHolder | CellRangeMetadata,
+        *,
+        row: int = -1,
+        column: int = -1,
+    ) -> Optional[CellType]:
+        if isinstance(definedName, str):
+            definedName = self._workbook.defined_names.get(definedName)
+            if definedName is None:
+                return None
+
+        if isinstance(definedName, DefinedName):
+            crh = self._getCellRange(definedName)
+            if crh is None:
+                return None
+            self._unused.discard(definedName)
+            stuff = crh
+        elif isinstance(definedName, (CellAndXBRLMetadataHolder, CellRangeMetadata)):
+            stuff = definedName
+        else:
+            return None
+
+        self._unused.discard(stuff.definedName)
+
+        cr = stuff.cellRange
+        ws = stuff.worksheet
+
+        if not all(
+            x is not None for x in (cr.min_row, cr.max_row, cr.min_col, cr.max_col)
+        ):
+            self._results.addMessage(
+                f"Named range {stuff.definedName.name} has an invalid cell range {cr.bounds}.",
+                Severity.ERROR,
+                MessageType.DevInfo,
+                excel_reference=excelDefinedNameRef(stuff.definedName),
+            )
+            return None
+
+        if cr.min_row == cr.max_row:
+            row = cr.min_row
+        if cr.min_col == cr.max_col:
+            column = cr.min_col
+
+        if row == -1:
+            row = cr.min_row
+        if column == -1:
+            column = cr.min_col
+
+        if not (cr.min_row <= row <= cr.max_row):
+            self._results.addMessage(
+                f"Row {row} has not been specified correctly.",
+                Severity.ERROR,
+                MessageType.DevInfo,
+                excel_reference=excelCellRangeRef(ws, cr),
+            )
+            row = cr.min_row
+        if not (cr.min_col <= column <= cr.max_col):
+            self._results.addMessage(
+                f"Column {column} has not been specified correctly.",
+                Severity.ERROR,
+                MessageType.DevInfo,
+                excel_reference=excelCellRangeRef(ws, cr),
+            )
+            column = cr.min_col
+
+        rows = list(
+            ws.iter_rows(min_row=row, max_row=row, min_col=column, max_col=column)
+        )
+        match len(rows):
+            case 0:
+                return None
+            case 1:
+                cells = rows[0]
+            case _:
+                return None
+
+        match len(cells):
+            case 0:
+                cell = None
+                self._results.addMessage(
+                    "No cells found in row of this named range.",
+                    Severity.ERROR,
+                    MessageType.DevInfo,
+                    excel_reference=excelCellRangeRef(ws, cr),
+                )
+            case 1:
+                cell = cells[0]
+            case _:
+                cell = None
+                self._results.addMessage(
+                    f"More than one cell found in range but only expected one cell. {cells}",
+                    Severity.ERROR,
+                    MessageType.DevInfo,
+                    excel_reference=excelCellRangeRef(ws, cr),
+                )
+
+        if cell is None or cell.value is None:
+            return None
+
+        if cell.value == EXCEL_PLACEHOLDER_VALUE:
+            self._results.addMessage(
+                f"Excel cell has an invalid stored value {EXCEL_PLACEHOLDER_VALUE}. Please check the Excel formula for this specific cell.",
+                Severity.ERROR,
+                MessageType.ExcelParsing,
+                excel_reference=excelCellRef(ws, cell),
+            )
+            return None
+        return cell
+
+    def getSingleValue(
+        self,
+        definedName: DefinedName | str,
+        *,
+        row: int = -1,
+        column: int = -1,
+    ) -> CellValueType:
+        if (
+            cell := self.getSingleCell(definedName, row=row, column=column)
+        ) is not None:
+            value = cell.value
+            if not isinstance(value, CellValueType):
+                value = str(value)
+            return value
+        return None
+
+    def getSingleStringValue(
+        self,
+        definedName: DefinedName | str,
+        *,
+        row: int = -1,
+        column: int = -1,
+        fallbackValue: str = "",
+    ) -> str:
+        value = self.getSingleValue(definedName, row=row, column=column)
+        return str(value) if value is not None else str(fallbackValue)
+
+    def getSingleDateValue(self, definedName: DefinedName | str) -> date:
+        value = self.getSingleValue(definedName)
+        return getDateFromValue(value)
