+++ tests/unitTests/xlsx_template_reader/conftest.py
@@ -1,0 +1,8 @@
+import pytest
+
+from mireport.taxonomy import loadBuiltInTaxonomyJSON
+
+
+@pytest.fixture(scope="session", autouse=True)
+def _load_taxonomies():
+    loadBuiltInTaxonomyJSON()
